Bugfixed. Double clicking on a song in the main table now works correctly
List .clear() apparently does not trigger the changelistener properly, and thus the queueMedia list was not properly cleaned during the call. Adding in a forced clear right afterwards fixed it. owever, this waranted the mediaplayer to be shut down earlier than intended, by neeing to stop it before the queueMedia list was cleared